# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "同频 API",
    "description": "同频 v0.1 API。浏览器端不会接收内部排序分数或原始模型特征。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/claims": {
      "get": {
        "tags": [
          "claims"
        ],
        "summary": "Get Claims",
        "operationId": "get_claims_api_v1_claims_get",
        "parameters": [
          {
            "name": "include_expired",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Expired"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClaimResponse"
                  },
                  "title": "Response Get Claims Api V1 Claims Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claim_id}/feedback": {
      "post": {
        "tags": [
          "claims"
        ],
        "summary": "Claim Feedback",
        "operationId": "claim_feedback_api_v1_claims__claim_id__feedback_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/claims/{claim_id}": {
      "patch": {
        "tags": [
          "claims"
        ],
        "summary": "Edit Claim",
        "operationId": "edit_claim_api_v1_claims__claim_id__patch",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "claims"
        ],
        "summary": "Remove Claim",
        "operationId": "remove_claim_api_v1_claims__claim_id__delete",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations": {
      "post": {
        "tags": [
          "matching"
        ],
        "summary": "Create Recommendations",
        "operationId": "create_recommendations_api_v1_recommendations_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationBundle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{candidate_id}": {
      "get": {
        "tags": [
          "matching"
        ],
        "summary": "Get Recommendation",
        "operationId": "get_recommendation_api_v1_recommendations__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateLead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{candidate_id}/actions": {
      "post": {
        "tags": [
          "matching"
        ],
        "summary": "Create Action",
        "operationId": "create_action_api_v1_recommendations__candidate_id__actions_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations": {
      "post": {
        "tags": [
          "interaction"
        ],
        "summary": "Create Invitation",
        "operationId": "create_invitation_api_v1_invitations_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/matches/{match_id}/messages": {
      "get": {
        "tags": [
          "interaction"
        ],
        "summary": "Get Messages",
        "operationId": "get_messages_api_v1_matches__match_id__messages_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Messages Api V1 Matches  Match Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "interaction"
        ],
        "summary": "Create Message",
        "operationId": "create_message_api_v1_matches__match_id__messages_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "more_like_this",
              "not_for_me",
              "saved",
              "seen"
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ActionRequest"
      },
...
      "CandidateLead": {
        "properties": {
          "recommendation_id": {
            "type": "string",
            "title": "Recommendation Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "common_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Common Signals"
          },
          "differences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Differences"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknowns"
          },
          "how_to_continue": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "How To Continue"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "type": "object",
        "required": [
          "recommendation_id",
          "candidate_id",
          "display_name",
          "headline",
          "common_signals",
          "differences",
          "unknowns",
          "how_to_continue",
          "evidence_ids"
        ],
        "title": "CandidateLead"
      },
      "ClaimFeedbackRequest": {
        "properties": {
          "feedback": {
            "type": "string",
            "enum": [
              "more_like_this",
              "not_like_this",
              "unsure"
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "feedback"
        ],
        "title": "ClaimFeedbackRequest"
      },
      "ClaimResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dimension": {
            "type": "string",
            "title": "Dimension"
          },
          "value": {
            "title": "Value"
          },
          "claim_type": {
            "type": "string",
            "title": "Claim Type"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "sensitivity": {
            "type": "string",
            "title": "Sensitivity"
          },
          "user_editable": {
            "type": "boolean",
            "title": "User Editable"
          },
          "user_confirmed": {
            "type": "boolean",
            "title": "User Confirmed"
          },
          "correction_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dimension",
          "value",
          "claim_type",
          "evidence_ids",
          "observed_at",
          "expires_at",
          "sensitivity",
          "user_editable",
          "user_confirmed",
          "correction_state"
        ],
        "title": "ClaimResponse"
      },
      "ClaimUpdateRequest": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "user_confirmed": {
            "type": "boolean",
            "title": "User Confirmed",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "ClaimUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationRequest": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "candidate_id"
        ],
        "title": "InvitationRequest"
      },
...
      "MatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "other_user": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Other User"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "other_user",
          "status",
          "created_at"
        ],
        "title": "MatchResponse"
      },
      "MessageCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body"
          },
          "client_message_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Message Id"
          }
        },
        "type": "object",
        "required": [
          "body",
          "client_message_id"
        ],
        "title": "MessageCreate"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "sender_id": {
            "type": "string",
            "title": "Sender Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "client_message_id": {
            "type": "string",
            "title": "Client Message Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "sender_id",
          "body",
          "client_message_id",
          "created_at",
          "read_at"
        ],
        "title": "MessageResponse"
      },
...
      "RecommendationBundle": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateLead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "session_id",
          "items"
        ],
        "title": "RecommendationBundle"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}